requirements.txt
# Core dependencies
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.2
python-dotenv==1.0.0
sqlalchemy==2.0.25
alembic==1.13.1
redis==5.0.1
httpx==0.26.0

# CrewAI and LLM
crewai==0.28.0
anthropic==0.15.0
openai==1.6.1

# WordPress integrations
python-wordpress-xmlrpc==2.3
requests==2.31.0

# SEO tools
semrush-api-client==0.1.5
pyahrefs==0.4.0

# Utilities
beautifulsoup4==4.12.2
lxml==4.9.3
python-slugify==8.0.1
Pillow==10.2.0
python-multipart==0.0.6
tenacity==8.2.3
